requirements.txt
pudb
pyyaml
jinja2
fastapi
uvicorn